Split ObjectEnumTests, add some more tests
